refactor: 구매 금액 최대 범위 확장 및 수익률 계산 방식 변경
구매금액을 long 범위까지 확장, 이에 따라서 수익률계산시 당첨금액의 합이 long 범위를 넘어가는 문제 해결

+++ src/main/java/edu/yk1028/LottoMachine.java
@@ -25,21 +25,21 @@ import edu.yk1028.Exception.MoneyLackException;
  *
  */
 public class LottoMachine {
-	public List<Lotto> cellLottos(int money) throws Exception {
+	public List<Lotto> cellLottos(long money) throws Exception {
 		if (money < LottoConstant.MIN_MONEY) {
 			throw new MoneyLackException();
 		}
 		return generateLottos(maximumNumberOfLotto(money));
 	}
 
-	private int maximumNumberOfLotto(int money) {
+	private long maximumNumberOfLotto(long money) {
 		return money / LottoConstant.LOTTO_PRICE;
 	}
 
-	private List<Lotto> generateLottos(int count) {
+	private List<Lotto> generateLottos(long count) {
 		List<Lotto> lottoList = new ArrayList<Lotto>();
 
-		for (int i = 0; i < count; i++) {
+		for (long i = 0; i < count; i++) {
 			lottoList.add(new Lotto(generateLottoNumbers()));
 		}
 		return lottoList;
